fix(jxl-encoder): near-lossless DC round-trip corruption at distance <= 0.02 (zenjxl#18)
Encoding at butteraugli distance <= 0.02 produced a broken round-trip: the
decoded image lost its low-frequency structure (~21-27 dB PSNR vs ~63 dB at
0.03) while the encoder spent MORE bits (larger file). Disambiguated with a
harness decoding each stream via BOTH zenjxl-decoder and jxl-oxide 0.12.6 --
both agreed the pre-fix 0.02 stream was garbage, so the bug is in the ENCODER,
not the decoder.

Root cause: quantized DC coefficients were stored as i16. The DC inv_factor
(INV_DC_QUANT[c]*scale_dc*dc_mul) grows as distance -> 0; at d <~ 0.025
legitimate DC exceeds +/-32767 and the `.round() as i16` cast saturated,
collapsing the DC plane.

Fix (two parts):
- Widen DC storage i16 -> i32 across the VarDCT DC path (transform, quantize_wp,
  dc_coding, dc_tree_learn, reconstruct, epf, bitstream, encoder, jpeg/encode,
  w44_audit dump). Matches the wire format (which already packs i32 DC
  residuals) and libjxl's internal i32 DC. In-range DC stays byte-identical.
- Clamp the VarDCT lossy distance up to a measured spec-conformant floor
  VARDCT_MIN_LOSSY_DISTANCE = 0.03. Below it the (now-correct) fine DC pushes
  the DC modular ANS stream past the JXL spec's 0x130000 ANS final-state check,
  so a conformant reference decoder (jxl-oxide) rejects the frame even though
  our decoder reconstructs it. 0.03 is the measured clean boundary across photo
  + high-contrast screen content; sub-floor requests emit conformant
  near-lossless output. LosslessConfig remains the path for bit-exact encoding.

Verified: frymire-srgb fixture (384px crop) round-trips 26.6 dB -> 63.2 dB at
distance 0.02 (zenjxl-decoder), oxide-conformant across all sub-0.03 distances,
normal distances (>= 0.05) byte-identical. Regression test
tests/nl_dc_roundtrip_issue18.rs asserts >= 40 dB + monotonicity (fails pre-fix
at 26.6 dB, passes post-fix). `cargo test` is currently blocked LOCALLY by
pre-existing zenjpeg 0.9.0-vs-0.8.4 dev-dep drift (unrelated to this fix, a
sibling repo); the test's fail-pre/pass-post was verified via the standalone
repro harness (identical encode->decode->PSNR path).

+++ jxl-encoder/src/jpeg/encode.rs
@@ -1524,7 +1524,7 @@ type WpDcTreeState = (
 /// `[jxl_channel][by][bx]` at that channel's native (subsampled) block
 /// resolution.
 type JpegCoefficientPlanes = (
-    [Vec<Vec<i16>>; 3],
+    [Vec<Vec<i32>>; 3],
     [Vec<Vec<[i32; BLOCK_SIZE]>>; 3],
     [Vec<Vec<u8>>; 3],
     [Vec<Vec<u16>>; 3],
@@ -1537,7 +1537,7 @@ fn map_jpeg_coefficients(
     budget: Option<&Arc<MemoryBudget>>,
 ) -> Result<JpegCoefficientPlanes> {
     let fallible = budget.is_some_and(|b| b.is_fallible());
-    let mut quant_dc: [Vec<Vec<i16>>; 3] = [Vec::new(), Vec::new(), Vec::new()];
+    let mut quant_dc: [Vec<Vec<i32>>; 3] = [Vec::new(), Vec::new(), Vec::new()];
     let mut quant_ac: [Vec<Vec<[i32; BLOCK_SIZE]>>; 3] = [Vec::new(), Vec::new(), Vec::new()];
     let mut nzeros: [Vec<Vec<u8>>; 3] = [Vec::new(), Vec::new(), Vec::new()];
     let mut raw_nzeros: [Vec<Vec<u16>>; 3] = [Vec::new(), Vec::new(), Vec::new()];
@@ -1580,10 +1580,13 @@ fn map_jpeg_coefficients(
                 // preserving the existing YCbCr/grayscale paths.
                 //
                 // The offset is small (~128 for typical luma DC quant
-                // = 8) and the result still fits in i16 because JPEG
-                // quantized DC is in [-1024, 1023] and the max offset
-                // is `1024 / 1` = 1024.
-                let dc = (comp.coeffs[base] as i32 + dc_offset[jxl_c]).clamp(-32768, 32767) as i16;
+                // = 8) and the result comfortably fits in i16 because
+                // JPEG quantized DC is in [-1024, 1023] and the max
+                // offset is `1024 / 1` = 1024. The quant_dc plane is i32
+                // now (issue #18 — VarDCT fine-distance DC overflowed
+                // i16), but JPEG-transcode DC keeps the same i16-range
+                // clamp so the transcode round-trip is byte-identical.
+                let dc = (comp.coeffs[base] as i32 + dc_offset[jxl_c]).clamp(-32768, 32767);
                 dc_row.push(dc);
 
                 // AC coefficients with transposition: JXL block[x*8+y] = JPEG[y*8+x]
